+++ ping/src/main/java/com/b6122/ping/config/SecurityConfig.java
@@ -2,7 +2,7 @@ package com.b6122.ping.config;
 
 import com.b6122.ping.config.jwt.JwtAuthorizationFilter;
 import com.b6122.ping.repository.datajpa.UserDataRepository;
-import com.b6122.ping.service.JwtService;
+import com.b6122.ping.config.jwt.JwtProvider;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -24,7 +24,7 @@ public class SecurityConfig {
     private final CorsConfig corsConfig;
     private final UserDetailsService userDetailsService;
     private final UserDataRepository userDataRepository;
-    private final JwtService jwtService;
+    private final JwtProvider jwtProvider;
 
     @Bean
     public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
@@ -41,7 +41,7 @@ public class SecurityConfig {
                 .sessionManagement((sessionManagement) -> sessionManagement.sessionCreationPolicy(SessionCreationPolicy.STATELESS))
                 .formLogin((formLogin) -> formLogin.disable())
 //                .addFilter(new JwtAuthenticationFilter((authenticationManager)))
-                .addFilter((new JwtAuthorizationFilter(authenticationManager, userDataRepository, jwtService)))
+                .addFilter((new JwtAuthorizationFilter(authenticationManager, userDataRepository, jwtProvider)))
                 .httpBasic((httpBasic) -> httpBasic.disable()) //Bearer 방식을 사용하기 위해 basic 인증 비활성화
                 .authorizeHttpRequests((authorize) ->
                         authorize
